+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ Unreleased
 ----------
 * BREAKING CHANGE: new argument for ``market_watch_plus``: ``best_limits_prepare_join``. To retain the old behaviour call this function with ``best_limits_prepare_join=False``.
 * Fixed occasional type-casting bug in ``market_watch_plus``.
+* Added easy-to-use default callbacks for ``MarketWatch`` class.
 
 v0.54.0
 -------